Move components to proper subpath
