+++ src/app/AppNotificationToast.tsx
@@ -1,7 +1,7 @@
 // Copyright (c) 2024 Tobias Briones. All rights reserved.
 // This file is part of https://github.com/mathswe/mathswe.com
 
-import NotificationToast from "@components/Toast.tsx";
+import NotificationToast from "@components/Toast/Toast.tsx";
 import { useAppDispatch, useAppSelector } from "@app/hooks.ts";
 import {
     hideNotificationToast,
