+++ src/components/CloseIcon/CloseIcon.tsx
